requirements.txt
# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# AI/ML Libraries
langchain>=0.1.0
langchain-aws>=0.1.0
opensearch-py>=2.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# CDK
aws-cdk-lib>=2.100.0
constructs>=10.0.0
